requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.6
django-anymail==13.1
django-celery-beat==2.8.1
django-channels==0.7.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.31.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
qrcode==8.2
redis==7.0.1
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
